refactor(frontend): split TipTap editor internals
