+++ apps/frontend/e2e/pages/RoomPage.ts
@@ -184,6 +184,7 @@ export class RoomPage {
     await this.messageInput.press('Enter');
     const message = this.getMessage(text);
     await expect(message.locator).toBeVisible({ timeout: TIMEOUTS.UI_FAST });
+    await this.waitForInputEditable();
     return message;
   }
 
@@ -197,6 +198,7 @@ export class RoomPage {
     await this.sendButton.click();
     const message = this.getMessage(text);
     await expect(message.locator).toBeVisible();
+    await this.waitForInputEditable();
     return message;
   }
 
@@ -689,7 +691,9 @@ export class RoomPage {
    * Assert that text is visible in the thread pane.
    */
   async expectTextInThreadPane(text: string): Promise<void> {
-    await expect(this.getThreadMessage(text).locator).toBeVisible({ timeout: TIMEOUTS.UI_STANDARD });
+    await expect(this.getThreadMessage(text).locator).toBeVisible({
+      timeout: TIMEOUTS.UI_STANDARD
+    });
   }
 
   /**
