+++ apps/frontend/src/lib/components/composer/ComposerToolbar.svelte
@@ -6,7 +6,7 @@
     ComposerFormattingCommand,
     ComposerFormattingState,
     TipTapEditorApi
-  } from './TipTapEditor.svelte';
+  } from './editorTypes';
 
   let {
     formattingState,
