+++ apps/frontend/src/lib/components/composer/TipTapEditor.svelte
@@ -18,863 +18,28 @@ and exposes a typed API for text manipulation (mentions, emoji, drafts).
 -->
 <script lang="ts">
   import { tick, untrack } from 'svelte';
-  import { Editor, Extension, InputRule, mergeAttributes, type JSONContent } from '@tiptap/core';
-  import {
-    Fragment,
-    Slice,
-    type Mark,
-    type Node as ProseMirrorNode,
-    type Schema
-  } from '@tiptap/pm/model';
-  import { Plugin, PluginKey, TextSelection } from '@tiptap/pm/state';
-  import StarterKit from '@tiptap/starter-kit';
-  import Link from '@tiptap/extension-link';
-  import CodeBlockLowlight from '@tiptap/extension-code-block-lowlight';
-  import { Markdown } from '@tiptap/markdown';
-  import Placeholder from '@tiptap/extension-placeholder';
-  import {
-    CODE_LANGUAGE_OPTIONS,
-    ensureCodeLanguagesLoaded,
-    lowlight
-  } from '$lib/codeHighlighting';
+  import { Editor } from '@tiptap/core';
+  import { Slice } from '@tiptap/pm/model';
+  import { CODE_LANGUAGE_OPTIONS, ensureCodeLanguagesLoaded } from '$lib/codeHighlighting';
   import * as m from '$lib/i18n/messages';
-  import type { QuoteInsertionContent, SelectedQuoteBlock } from '$lib/state/room';
-
-  const markdownLinkInputRegex = /(^|\s)\[([^\]\n]+)\]\((https?:\/\/[^\s)]+)\)$/;
-  const markdownLinkPasteRegex = /\[([^\]\n]+)\]\((https?:\/\/[^\s)]+)\)/g;
-  const codeFenceLineRegex = /^```([\w-]+)?$/;
-  const markdownBulletListLineRegex = /^[ \t]{0,3}[-+*]\s(.*)$/;
-  const markdownOrderedListLineRegex = /^[ \t]{0,3}(\d{1,9})[.)]\s(.*)$/;
-
-  const ComposerCodeBlockLowlight = CodeBlockLowlight.extend({
-    renderHTML({ node, HTMLAttributes }) {
-      const language = node.attrs.language || 'text';
-
-      return [
-        'pre',
-        mergeAttributes(this.options.HTMLAttributes, HTMLAttributes, {
-          'data-language': language
-        }),
-        [
-          'code',
-          {
-            class: language ? this.options.languageClassPrefix + language : null
-          },
-          0
-        ]
-      ];
-    }
-  });
-
-  const ComposerLink = Link.extend({ inclusive: false });
-
-  function paragraphTextWithLineBreaks(node: ProseMirrorNode) {
-    return node.textBetween(0, node.content.size, '\n', '\n');
-  }
-
-  function isDefaultEmptyDocument(doc: ProseMirrorNode): boolean {
-    if (doc.childCount !== 1) return false;
-    const firstChild = doc.firstChild;
-    return firstChild?.type.name === 'paragraph' && firstChild.content.size === 0;
-  }
-
-  function createParagraphFromText(schema: Schema, text: string) {
-    const paragraph = schema.nodes.paragraph;
-    const hardBreak = schema.nodes.hardBreak;
-    if (!text) return paragraph.create();
-
-    const content = text.split('\n').flatMap((line, index, lines) => {
-      const nodes = [];
-      if (line) nodes.push(schema.text(line));
-      if (index < lines.length - 1 && hardBreak) nodes.push(hardBreak.create());
-      return nodes;
-    });
-
-    return paragraph.create(null, content);
-  }
-
-  function buildCodeFenceReplacement({
-    schema,
-    paragraph,
-    openLineIndex,
-    closeLineIndex,
-    appendTrailingParagraph
-  }: {
-    schema: Schema;
-    paragraph: Parameters<typeof paragraphTextWithLineBreaks>[0];
-    openLineIndex: number;
-    closeLineIndex: number | null;
-    appendTrailingParagraph: boolean;
-  }) {
-    const text = paragraphTextWithLineBreaks(paragraph);
-    const lines = text.split('\n');
-    const openingMatch = lines[openLineIndex]?.match(codeFenceLineRegex);
-    const codeBlock = schema.nodes.codeBlock;
-    if (!openingMatch || !codeBlock) return null;
-
-    const beforeText = lines.slice(0, openLineIndex).join('\n');
-    const codeText =
-      closeLineIndex === null ? '' : lines.slice(openLineIndex + 1, closeLineIndex).join('\n');
-    const afterText =
-      closeLineIndex === null
-        ? lines.slice(openLineIndex + 1).join('\n')
-        : lines.slice(closeLineIndex + 1).join('\n');
-    const language = openingMatch[1] || null;
-
-    const nodes = [];
-    const beforeNode = beforeText ? createParagraphFromText(schema, beforeText) : null;
-    if (beforeNode) nodes.push(beforeNode);
-    const codeNode = codeBlock.create(
-      language ? { language } : undefined,
-      codeText ? schema.text(codeText) : null
-    );
-    nodes.push(codeNode);
-    if (afterText) {
-      nodes.push(createParagraphFromText(schema, afterText));
-    } else if (appendTrailingParagraph) {
-      nodes.push(schema.nodes.paragraph.create());
-    }
-
-    return { nodes, codeNode, beforeNodeSize: beforeNode?.nodeSize ?? 0 };
-  }
-
-  function buildListMarkerReplacement({
-    schema,
-    paragraph,
-    markerLineIndex
-  }: {
-    schema: Schema;
-    paragraph: Parameters<typeof paragraphTextWithLineBreaks>[0];
-    markerLineIndex: number;
-  }) {
-    const text = paragraphTextWithLineBreaks(paragraph);
-    const lines = text.split('\n');
-    const markerLine = lines[markerLineIndex] ?? '';
-    const bulletMatch = markerLine.match(markdownBulletListLineRegex);
-    const orderedMatch = markerLine.match(markdownOrderedListLineRegex);
-    if (!bulletMatch && !orderedMatch) return null;
-
-    const listNodeType = bulletMatch ? schema.nodes.bulletList : schema.nodes.orderedList;
-    const listItem = schema.nodes.listItem;
-    const paragraphNode = schema.nodes.paragraph;
-    if (!listNodeType || !listItem || !paragraphNode) return null;
-
-    const itemText = bulletMatch ? (bulletMatch[1] ?? '') : (orderedMatch?.[2] ?? '');
-    const beforeText = lines.slice(0, markerLineIndex).join('\n');
-    const afterText = lines.slice(markerLineIndex + 1).join('\n');
-    const beforeNode = beforeText ? createParagraphFromText(schema, beforeText) : null;
-    const itemParagraph = itemText
-      ? paragraphNode.create(null, schema.text(itemText))
-      : paragraphNode.create();
-    const listAttrs = orderedMatch ? { start: Number.parseInt(orderedMatch[1], 10) } : undefined;
-    const listNode = listNodeType.create(listAttrs, listItem.create(null, itemParagraph));
-    const nodes = [];
-
-    if (beforeNode) nodes.push(beforeNode);
-    nodes.push(listNode);
-    if (afterText) nodes.push(createParagraphFromText(schema, afterText));
-
-    return {
-      nodes,
-      itemText,
-      listNode,
-      beforeNodeSize: beforeNode?.nodeSize ?? 0
-    };
-  }
-
-  const MarkdownLinkInputRule = Extension.create({
-    name: 'markdownLinkInputRule',
-
-    addInputRules() {
-      return [
-        new InputRule({
-          find: markdownLinkInputRegex,
-          handler: ({ state, range, match }) => {
-            const prefix = match[1] ?? '';
-            const label = match[2];
-            const href = match[3];
-            const linkType = state.schema.marks.link;
-            if (!label || !href || !linkType) return null;
-
-            const from = range.from + prefix.length;
-            const to = range.to;
-            const tr = state.tr;
-
-            tr.delete(from, to);
-            tr.insertText(label, from);
-            tr.addMark(from, from + label.length, linkType.create({ href }));
-            tr.removeStoredMark(linkType);
-          }
-        })
-      ];
-    }
-  });
-
-  const CompletedMarkdownCodeFence = Extension.create({
-    name: 'completedMarkdownCodeFence',
-
-    addProseMirrorPlugins() {
-      return [
-        new Plugin({
-          key: new PluginKey('completedMarkdownCodeFence'),
-          appendTransaction: (transactions, _oldState, newState) => {
-            if (!transactions.some((transaction) => transaction.docChanged)) return null;
-
-            let paragraphPos = 0;
-            for (let index = 0; index < newState.doc.childCount; index += 1) {
-              const paragraph = newState.doc.child(index);
-              const currentParagraphPos = paragraphPos;
-              paragraphPos += paragraph.nodeSize;
-              if (paragraph.type.name !== 'paragraph') continue;
-
-              const lines = paragraphTextWithLineBreaks(paragraph).split('\n');
-              for (let openLineIndex = 0; openLineIndex < lines.length; openLineIndex += 1) {
-                if (!codeFenceLineRegex.test(lines[openLineIndex] ?? '')) continue;
-
-                for (
-                  let closeLineIndex = openLineIndex + 1;
-                  closeLineIndex < lines.length;
-                  closeLineIndex += 1
-                ) {
-                  if (lines[closeLineIndex] !== '```') continue;
-
-                  const appendTrailingParagraph =
-                    currentParagraphPos + paragraph.nodeSize === newState.doc.content.size &&
-                    closeLineIndex === lines.length - 1;
-                  const replacement = buildCodeFenceReplacement({
-                    schema: newState.schema,
-                    paragraph,
-                    openLineIndex,
-                    closeLineIndex,
-                    appendTrailingParagraph
-                  });
-                  if (!replacement) return null;
-
-                  const tr = newState.tr.replaceWith(
-                    currentParagraphPos,
-                    currentParagraphPos + paragraph.nodeSize,
-                    replacement.nodes
-                  );
-                  const codeEnd =
-                    currentParagraphPos +
-                    replacement.beforeNodeSize +
-                    replacement.codeNode.nodeSize;
-                  tr.setSelection(TextSelection.near(tr.doc.resolve(codeEnd + 1), 1));
-                  return tr;
-                }
-              }
-            }
-
-            return null;
-          }
-        })
-      ];
-    }
-  });
-
-  const MarkdownListMarkerAfterHardBreak = Extension.create({
-    name: 'markdownListMarkerAfterHardBreak',
-
-    addProseMirrorPlugins() {
-      return [
-        new Plugin({
-          key: new PluginKey('markdownListMarkerAfterHardBreak'),
-          appendTransaction: (transactions, _oldState, newState) => {
-            if (!transactions.some((transaction) => transaction.docChanged)) return null;
-
-            const selectionFrom = newState.selection.$from;
-            if (selectionFrom.depth !== 1 || selectionFrom.parent.type.name !== 'paragraph') {
-              return null;
-            }
-
-            const paragraph = selectionFrom.parent;
-            const paragraphPos = selectionFrom.before(1);
-            const textBeforeCursor = paragraph.textBetween(
-              0,
-              selectionFrom.parentOffset,
-              '\n',
-              '\n'
-            );
-            const currentLineIndex = textBeforeCursor.split('\n').length - 1;
-            const currentLine = textBeforeCursor.split('\n').at(-1) ?? '';
-
-            if (
-              !markdownBulletListLineRegex.test(currentLine) &&
-              !markdownOrderedListLineRegex.test(currentLine)
-            ) {
-              return null;
-            }
-
-            const replacement = buildListMarkerReplacement({
-              schema: newState.schema,
-              paragraph,
-              markerLineIndex: currentLineIndex
-            });
-            if (!replacement) return null;
-
-            const listPosition = paragraphPos + replacement.beforeNodeSize;
-            const selectionPosition = listPosition + 3 + replacement.itemText.length;
-            const tr = newState.tr.replaceWith(
-              paragraphPos,
-              paragraphPos + paragraph.nodeSize,
-              replacement.nodes
-            );
-            tr.setSelection(TextSelection.create(tr.doc, selectionPosition));
-            return tr;
-          }
-        })
-      ];
-    }
-  });
-
-  const TrailingParagraphAfterCodeBlock = Extension.create({
-    name: 'trailingParagraphAfterCodeBlock',
-
-    addProseMirrorPlugins() {
-      return [
-        new Plugin({
-          key: new PluginKey('trailingParagraphAfterCodeBlock'),
-          appendTransaction: (_transactions, _oldState, newState) => {
-            const paragraph = newState.schema.nodes.paragraph;
-            const lastChild = newState.doc.lastChild;
-            if (!paragraph || !lastChild || lastChild.type.name !== 'codeBlock') return null;
-
-            return newState.tr.insert(newState.doc.content.size, paragraph.create());
-          }
-        })
-      ];
-    }
-  });
-
-  const ClearMarksOnEmptyDocument = Extension.create({
-    name: 'clearMarksOnEmptyDocument',
-
-    addProseMirrorPlugins() {
-      return [
-        new Plugin({
-          key: new PluginKey('clearMarksOnEmptyDocument'),
-          appendTransaction: (transactions, _oldState, newState) => {
-            if (!transactions.some((transaction) => transaction.docChanged)) return null;
-            if (!isDefaultEmptyDocument(newState.doc)) return null;
-            if ((newState.storedMarks?.length ?? 0) === 0) return null;
-
-            return newState.tr.setStoredMarks([]);
-          }
-        })
-      ];
-    }
-  });
-
-  function encodeMarkdownTextHtml(text: string): string {
-    return text.replace(/&/g, '&amp;').replace(/</g, '&lt;');
-  }
-
-  function decodeSerializedTextEntities(text: string): string {
-    return text
-      .split(/(\n)/)
-      .map((part) => {
-        if (part === '\n') return part;
-
-        const leadingBlockquoteMarker = part.match(/^( {0,3})&gt;(?=\s|$)/);
-        const protectedPart = leadingBlockquoteMarker
-          ? `${leadingBlockquoteMarker[1]}__CHATTO_LITERAL_BLOCKQUOTE_MARKER__${part.slice(leadingBlockquoteMarker[0].length)}`
-          : part;
-
-        return protectedPart
-          .replace(/&lt;/g, '<')
-          .replace(/&gt;/g, '>')
-          .replace(/&amp;/g, '&')
-          .replace('__CHATTO_LITERAL_BLOCKQUOTE_MARKER__', '&gt;');
-      })
-      .join('');
-  }
-
-  function transformOutsideMarkdownLinkDestinations(
-    text: string,
-    transformText: (text: string) => string
-  ): string {
-    let result = '';
-    let index = 0;
-    let textStart = 0;
-    const bracketStack: number[] = [];
-
-    while (index < text.length) {
-      const char = text[index];
-      if (char === '\\') {
-        index += 2;
-        continue;
-      }
-
-      if (char === '[') {
-        bracketStack.push(index);
-        index += 1;
-        continue;
-      }
-
-      if (char !== ']' || text[index + 1] !== '(' || bracketStack.length === 0) {
-        index += 1;
-        continue;
-      }
-
-      bracketStack.pop();
-      const destinationStart = index;
-      const destinationContentStart = destinationStart + 2;
-      let destinationEnd = destinationContentStart;
-      let nestedParens = 0;
-      while (destinationEnd < text.length) {
-        const char = text[destinationEnd];
-        if (char === '\\') {
-          destinationEnd += 2;
-          continue;
-        }
-        if (char === '(') {
-          nestedParens += 1;
-        } else if (char === ')') {
-          if (nestedParens === 0) break;
-          nestedParens -= 1;
-        }
-        destinationEnd += 1;
-      }
-
-      if (destinationEnd >= text.length) {
-        result += transformOutsideMarkdownAutolinks(text.slice(textStart), transformText);
-        return result;
-      }
-
-      result += transformOutsideMarkdownAutolinks(
-        text.slice(textStart, destinationContentStart),
-        transformText
-      );
-      result += text.slice(destinationContentStart, destinationEnd + 1);
-      index = destinationEnd + 1;
-      textStart = index;
-    }
-
-    result += transformOutsideMarkdownAutolinks(text.slice(textStart), transformText);
-    return result;
-  }
-
-  function transformOutsideMarkdownAutolinks(
-    text: string,
-    transformText: (text: string) => string
-  ): string {
-    let result = '';
-    let index = 0;
-    const autolinkPattern = /<https?:\/\/[^\s<>]+>/gi;
-
-    for (const match of text.matchAll(autolinkPattern)) {
-      result += transformText(text.slice(index, match.index));
-      result += match[0];
-      index = match.index + match[0].length;
-    }
-
-    result += transformText(text.slice(index));
-    return result;
-  }
-
-  type MarkdownTransformOptions = {
-    skipLinkDestinations?: boolean;
-    preserveInlineCode?: boolean;
-  };
-
-  function transformMarkdownTextSegment(
-    text: string,
-    transformText: (text: string) => string,
-    { skipLinkDestinations = false }: MarkdownTransformOptions = {}
-  ): string {
-    return skipLinkDestinations
-      ? transformOutsideMarkdownLinkDestinations(text, transformText)
-      : transformText(text);
-  }
-
-  function transformOutsideInlineCode(
-    line: string,
-    transformText: (text: string) => string,
-    options: MarkdownTransformOptions = {}
-  ): string {
-    let result = '';
-    let index = 0;
-
-    while (index < line.length) {
-      const codeStart = line.indexOf('`', index);
-      if (codeStart === -1) {
-        result += transformMarkdownTextSegment(line.slice(index), transformText, options);
-        break;
-      }
-
-      result += transformMarkdownTextSegment(line.slice(index, codeStart), transformText, options);
-
-      let delimiterEnd = codeStart + 1;
-      while (line[delimiterEnd] === '`') delimiterEnd += 1;
-
-      const delimiter = line.slice(codeStart, delimiterEnd);
-      const codeEnd = line.indexOf(delimiter, delimiterEnd);
-      if (codeEnd === -1) {
-        result += transformMarkdownTextSegment(line.slice(codeStart), transformText, options);
-        break;
-      }
-
-      result += line.slice(codeStart, codeEnd + delimiter.length);
-      index = codeEnd + delimiter.length;
-    }
-
-    return result;
-  }
-
-  function transformMarkdownOutsideCode(
-    markdown: string,
-    transformText: (text: string) => string,
-    options: MarkdownTransformOptions = {}
-  ): string {
-    const lines = markdown.match(/[^\n]*(?:\n|$)/g) ?? [];
-    if (lines[lines.length - 1] === '') {
-      lines.pop();
-    }
-
-    let result = '';
-    let pendingText = '';
-    let inFence = false;
-    let fenceChar = '';
-    let fenceLength = 0;
-    let canStartIndentedCode = true;
-
-    const flushPendingText = () => {
-      if (!pendingText) return;
-      result +=
-        options.preserveInlineCode === false
-          ? transformMarkdownTextSegment(pendingText, transformText, options)
-          : transformOutsideInlineCode(pendingText, transformText, options);
-      pendingText = '';
-    };
-
-    for (const lineWithBreak of lines) {
-      const hasLineBreak = lineWithBreak.endsWith('\n');
-      const line = hasLineBreak ? lineWithBreak.slice(0, -1) : lineWithBreak;
-      const blockquoteContent = line.replace(/^(?: {0,3}> ?)+/, '');
-
-      if (/^ *$/.test(blockquoteContent)) {
-        if (inFence) {
-          result += lineWithBreak;
-        } else {
-          pendingText += lineWithBreak;
-        }
-        canStartIndentedCode = true;
-        continue;
-      }
-
-      const fence = blockquoteContent.match(/^ {0,3}(`{3,}|~{3,})/);
-      if (fence) {
-        flushPendingText();
-        const marker = fence[1];
-        if (!inFence) {
-          inFence = true;
-          fenceChar = marker[0];
-          fenceLength = marker.length;
-        } else if (
-          marker[0] === fenceChar &&
-          marker.length >= fenceLength &&
-          new RegExp(`^ {0,3}\\${fenceChar}{${fenceLength},} *$`).test(blockquoteContent)
-        ) {
-          inFence = false;
-          fenceChar = '';
-          fenceLength = 0;
-        }
-
-        result += lineWithBreak;
-        canStartIndentedCode = true;
-        continue;
-      }
-
-      if (inFence) {
-        result += lineWithBreak;
-        continue;
-      }
-
-      if (canStartIndentedCode && /^( {4,}|\t)/.test(blockquoteContent)) {
-        flushPendingText();
-        result += lineWithBreak;
-        continue;
-      }
-
-      pendingText += lineWithBreak;
-      canStartIndentedCode = false;
-    }
-
-    flushPendingText();
-    return result;
-  }
-
-  function escapeMarkdownHtml(markdown: string): string {
-    return transformMarkdownOutsideCode(markdown, encodeMarkdownTextHtml, {
-      skipLinkDestinations: true
-    });
-  }
-
-  function hasUnescapedPipe(line: string): boolean {
-    for (let index = 0; index < line.length; index++) {
-      if (line[index] === '|' && line[index - 1] !== '\\') return true;
-    }
-    return false;
-  }
-
-  function isGfmTableDelimiter(line: string): boolean {
-    const content = line.replace(/^(?: {0,3}> ?)+/, '').trim();
-    const cells = content.split('|');
-    if (cells[0]?.trim() === '') cells.shift();
-    if (cells.at(-1)?.trim() === '') cells.pop();
-    return cells.length > 0 && cells.every((cell) => /^:?-+:?$/.test(cell.trim()));
-  }
-
-  function escapeGfmTablesForEditor(markdown: string): string {
-    return transformMarkdownOutsideCode(
-      markdown,
-      (text) => {
-        const lines = text.split('\n');
-        for (let index = 1; index < lines.length; index++) {
-          const header = lines[index - 1].replace(/^(?: {0,3}> ?)+/, '').trim();
-          if (!hasUnescapedPipe(header) || !isGfmTableDelimiter(lines[index])) continue;
-
-          // The composer intentionally has no table node. Keep the source as
-          // editable prose by making the delimiter invisible to TipTap's GFM
-          // parser. The word joiner is invisible and is removed again when
-          // serializing, so the delimiter remains visually unchanged.
-          lines[index] = lines[index].replace('-', '-\u2060');
-        }
-        return lines.join('\n');
-      },
-      { preserveInlineCode: false }
-    );
-  }
-
-  function prepareMarkdownForEditor(markdown: string): string {
-    return escapeGfmTablesForEditor(escapeMarkdownHtml(markdown));
-  }
-
-  function decodeSerializedMarkdownText(markdown: string): string {
-    return transformMarkdownOutsideCode(markdown, decodeSerializedTextEntities);
-  }
-
-  function hasTrailingEmptyParagraph(e: Editor): boolean {
-    if (e.state.doc.childCount <= 1) return false;
-    const lastChild = e.state.doc.lastChild;
-    return lastChild?.type.name === 'paragraph' && lastChild.content.size === 0;
-  }
-
-  function trimSerializedTrailingEmptyParagraph(markdown: string, e: Editor): string {
-    if (!hasTrailingEmptyParagraph(e)) return markdown;
-    return markdown.replace(/(?:\n\n(?:&nbsp;|\u00a0))+$/, '');
-  }
-
-  function normalizeSerializedHardBreaksBeforeLists(markdown: string): string {
-    return markdown.replace(/ {2,}(\n\s*\n\s*(?:[-+*]|\d{1,9}[.)])\s)/g, '$1');
-  }
-
-  function normalizeSerializedGfmTableHardBreaks(markdown: string): string {
-    return transformMarkdownOutsideCode(
-      markdown,
-      (text) => {
-        const lines = text.split('\n');
-        for (let index = 1; index < lines.length; index++) {
-          const header = lines[index - 1].replace(/ {2,}$/, '');
-          const delimiter = lines[index].replace(/ {2,}$/, '');
-          const unescapedDelimiter = delimiter.replace('\u2060', '');
-          if (!hasUnescapedPipe(header) || !isGfmTableDelimiter(unescapedDelimiter)) continue;
-
-          lines[index - 1] = header;
-          lines[index] = unescapedDelimiter;
-          for (let rowIndex = index + 1; rowIndex < lines.length; rowIndex++) {
-            const row = lines[rowIndex].replace(/ {2,}$/, '');
-            if (!hasUnescapedPipe(row)) break;
-            lines[rowIndex] = row;
-          }
-        }
-        return lines.join('\n');
-      },
-      { preserveInlineCode: false }
-    );
-  }
-
-  function encodeSerializedHeadingClosingHashes(markdown: string): string {
-    return transformMarkdownOutsideCode(
-      markdown,
-      (text) =>
-        text.replace(
-          /(^|\n)((?:[ \t]{0,3}>[ \t]?)*[ \t]{0,3}#{1,6}[ \t]+[^\n]*?)([ \t]+)(#{1,})([ \t]*)(?=\n|$)/g,
-          (_match, lineStart, headingStart, separator, hashes, trailingWhitespace) =>
-            `${lineStart}${headingStart}${separator}${hashes.replace(/#/g, '&#35;')}${trailingWhitespace}`
-        ),
-      { preserveInlineCode: false }
-    );
-  }
-
-  function getSerializedMarkdown(e: Editor): string {
-    return normalizeSerializedHardBreaksBeforeLists(
-      normalizeSerializedGfmTableHardBreaks(
-        encodeSerializedHeadingClosingHashes(
-          trimSerializedTrailingEmptyParagraph(decodeSerializedMarkdownText(e.getMarkdown()), e)
-        )
-      )
-    );
-  }
-
-  function applyDestinationMarks(node: ProseMirrorNode, marks: readonly Mark[]): ProseMirrorNode {
-    const children: ProseMirrorNode[] = [];
-
-    node.forEach((child) => {
-      if (child.isText) {
-        const combinedMarks = marks.reduce((current, mark) => mark.addToSet(current), child.marks);
-        children.push(child.mark(node.type.allowedMarks(combinedMarks)));
-      } else {
-        children.push(applyDestinationMarks(child, marks));
-      }
-    });
-
-    return node.copy(Fragment.fromArray(children));
-  }
-
-  function createClipboardInlineContent(
-    text: string,
-    schema: Schema,
-    destinationMarks: readonly Mark[]
-  ): ProseMirrorNode[] {
-    const linkType = schema.marks.link;
-    if (!text || !linkType) return text ? [schema.text(text, destinationMarks)] : [];
-
-    const nodes: ProseMirrorNode[] = [];
-    const nonLinkMarks = destinationMarks.filter((mark) => mark.type !== linkType);
-    let index = 0;
-
-    for (const match of text.matchAll(markdownLinkPasteRegex)) {
-      const matchIndex = match.index;
-      const label = match[1];
-      const href = match[2];
-      if (!label || !href || text[matchIndex - 1] === '!' || text[matchIndex - 1] === '\\') {
-        continue;
-      }
-
-      if (matchIndex > index)
-        nodes.push(schema.text(text.slice(index, matchIndex), destinationMarks));
-      nodes.push(schema.text(label, [...nonLinkMarks, linkType.create({ href })]));
-      index = matchIndex + match[0].length;
-    }
-
-    if (index < text.length) nodes.push(schema.text(text.slice(index), destinationMarks));
-    return nodes;
-  }
-
-  function createClipboardContent(
-    text: string,
-    schema: Schema,
-    destinationMarks: readonly Mark[]
-  ): Fragment {
-    const paragraphType = schema.nodes.paragraph;
-    const hardBreakType = schema.nodes.hardBreak;
-    const paragraphMarks = paragraphType.allowedMarks(destinationMarks);
-
-    return Fragment.fromArray(
-      text.split(/\n{2,}/).map((paragraphText) => {
-        const inlineNodes: ProseMirrorNode[] = [];
-        const lines = paragraphText.split('\n');
-
-        lines.forEach((line, index) => {
-          inlineNodes.push(...createClipboardInlineContent(line, schema, paragraphMarks));
-          if (index < lines.length - 1) inlineNodes.push(hardBreakType.create());
-        });
-
-        return paragraphType.create(null, inlineNodes);
-      })
-    );
-  }
-
-  function hasDefaultEmptyDocument(e: Editor): boolean {
-    return isDefaultEmptyDocument(e.state.doc);
-  }
-
-  function isSelectionInTrailingEmptyParagraph(e: Editor): boolean {
-    const { doc, selection } = e.state;
-    if (!selection.empty || doc.childCount <= 1) return false;
-
-    const selectionFrom = selection.$from;
-    if (selectionFrom.depth !== 1 || selectionFrom.parent.type.name !== 'paragraph') return false;
-    if (selectionFrom.parent.content.size !== 0 || selectionFrom.parentOffset !== 0) return false;
-    if (selectionFrom.after(1) !== doc.content.size) return false;
-
-    const previousNode = doc.child(doc.childCount - 2);
-    return previousNode.type.name !== 'paragraph' || previousNode.content.size > 0;
-  }
-
-  function hasRichStructure(e: Editor): boolean {
-    let found = false;
-    e.state.doc.descendants((node) => {
-      if (
-        ['heading', 'bulletList', 'orderedList', 'blockquote', 'codeBlock'].includes(node.type.name)
-      ) {
-        found = true;
-        return false;
-      }
-      return true;
-    });
-    return found;
-  }
-
-  function quoteParagraphsForText(text: string): JSONContent[] {
-    return text
-      .replace(/\r\n?/g, '\n')
-      .split('\n')
-      .map((line) => ({
-        type: 'paragraph',
-        content: line ? [{ type: 'text', text: line }] : undefined
-      }));
-  }
-
-  function normalizeQuoteInsertionContent(text: QuoteInsertionContent): SelectedQuoteBlock[] {
-    if (typeof text !== 'string') {
-      return text
-        .map((block) => ({
-          quoteDepth: Math.max(0, Math.floor(block.quoteDepth)),
-          text: block.text.replace(/\r\n?/g, '\n').trim()
-        }))
-        .filter((block) => block.text.length > 0);
-    }
-
-    const normalized = text.replace(/\r\n?/g, '\n').trim();
-    if (!normalized) return [];
-    return normalized.split('\n').map((line) => ({ quoteDepth: 0, text: line }));
-  }
-
-  function buildQuoteContent(blocks: SelectedQuoteBlock[]): JSONContent[] {
-    const content: JSONContent[] = [];
-
-    for (const block of blocks) {
-      let target = content;
-
-      for (let depth = 0; depth < block.quoteDepth; depth++) {
-        let current = target.at(-1);
-        if (current?.type !== 'blockquote') {
-          current = { type: 'blockquote', content: [] };
-          target.push(current);
-        }
-        current.content ??= [];
-        target = current.content;
-      }
-
-      target.push(...quoteParagraphsForText(block.text));
-    }
-
-    return content;
-  }
-
-  export type ComposerFormattingCommand =
-    | 'bold'
-    | 'italic'
-    | 'inlineCode'
-    | 'heading'
-    | 'bulletList'
-    | 'orderedList'
-    | 'blockquote'
-    | 'codeBlock';
-
-  export type ComposerFormattingState = Record<ComposerFormattingCommand, boolean>;
+  import type { QuoteInsertionContent } from '$lib/state/room';
+  import type {
+    ComposerFormattingCommand,
+    ComposerFormattingState,
+    TipTapEditorApi
+  } from './editorTypes';
+  import { createComposerExtensions } from './extensions';
+  import {
+    applyDestinationMarks,
+    buildQuoteContent,
+    createClipboardContent,
+    getSerializedMarkdown,
+    hasDefaultEmptyDocument,
+    hasRichStructure,
+    isSelectionInTrailingEmptyParagraph,
+    normalizeQuoteInsertionContent,
+    prepareMarkdownForEditor
+  } from './markdown';
 
   const emptyFormattingState: ComposerFormattingState = {
     bold: false,
@@ -885,33 +50,6 @@ and exposes a typed API for text manipulation (mentions, emoji, drafts).
     orderedList: false,
     blockquote: false,
     codeBlock: false
-  };
-
-  export type TipTapEditorApi = {
-    /** Get the editor's plain text content */
-    getText: () => string;
-    /** Set editor content from markdown */
-    setContent: (markdown: string) => void;
-    /** Focus the editor */
-    focus: (position?: 'start' | 'end') => void;
-    /** Get plain text from document start to cursor position */
-    getTextBeforeCursor: () => string;
-    /** Whether the current selection is inside a code block */
-    isInCodeBlock: () => boolean;
-    /**
-     * Replace N characters before the cursor with new text.
-     * Used for mention/emoji completion where we know the pattern
-     * length relative to the cursor.
-     */
-    replaceTextBeforeCursor: (charCount: number, replacement: string) => void;
-    /** Insert plain text at the current cursor position. */
-    insertText: (text: string) => void;
-    /** Toggle a Markdown formatting command at the current selection. */
-    toggleFormatting: (command: ComposerFormattingCommand) => void;
-    /** Insert selected reply text as a blockquote at the current cursor. */
-    insertQuote: (text: QuoteInsertionContent) => void;
-    /** Insert the same block break the editor would create for a plain Enter key. */
-    insertBlockBreak: () => void;
   };
 
   let {
@@ -1322,35 +460,7 @@ and exposes a typed API for text manipulation (mentions, emoji, drafts).
       () =>
         new Editor({
           element: editorElement,
-          extensions: [
-            StarterKit.configure({
-              // Keep the composer subset aligned with the rendered message markdown.
-              codeBlock: false,
-              strike: false,
-              underline: false,
-              horizontalRule: false,
-              trailingNode: false,
-              link: false
-            }),
-            ComposerLink.configure({
-              autolink: true,
-              linkOnPaste: true,
-              openOnClick: false,
-              enableClickSelection: true
-            }),
-            Markdown.configure({
-              markedOptions: {
-                breaks: true
-              }
-            }),
-            ComposerCodeBlockLowlight.configure({ lowlight }),
-            MarkdownLinkInputRule,
-            CompletedMarkdownCodeFence,
-            MarkdownListMarkerAfterHardBreak,
-            TrailingParagraphAfterCodeBlock,
-            ClearMarksOnEmptyDocument,
-            Placeholder.configure({ placeholder })
-          ],
+          extensions: createComposerExtensions(placeholder),
           content: '',
           contentType: 'markdown',
           editable,
@@ -1458,6 +568,7 @@ and exposes a typed API for text manipulation (mentions, emoji, drafts).
     <div class="flex min-w-0 flex-wrap items-center gap-1.5 text-xs text-muted">
       <div class="flex min-w-0 items-center gap-1">
         <input
+          name="composer-link-url"
           aria-label={m['composer.link_url']()}
           title={m['composer.link_url']()}
           value={linkHrefDraft}
@@ -1513,6 +624,7 @@ and exposes a typed API for text manipulation (mentions, emoji, drafts).
         <span>{activeCodeBlockLanguageLabel}</span>
         <span class="iconify size-3 uil--angle-down"></span>
         <select
+          name="composer-code-language"
           aria-label={m['composer.code_language']()}
           title={m['composer.code_language']()}
           value={activeCodeBlockLanguage}
